Move assert_eq to testing submodule
This way, importing it won't require pytest.

- Closes #16062
