+++ python/cudf/cudf/testing/__init__.py
@@ -1,7 +1,9 @@
-# Copyright (c) 2020, NVIDIA CORPORATION.
+# Copyright (c) 2020-2024, NVIDIA CORPORATION.
 
 from cudf.testing.testing import (
+    assert_eq,
     assert_frame_equal,
     assert_index_equal,
+    assert_neq,
     assert_series_equal,
 )
